+++ two-pointers/1358.number-of-substrings-containing-all-three-characters.cpp
@@ -3,10 +3,10 @@
 class Solution {
 public:
   int numberOfSubstrings(std::string s) {
-    int j = 0;
+    size_t j = 0;
     int count1 = 0, count2 = 0, count3 = 0;
     int ret = 0;
-    for (int i = 0; i < s.size(); i++) {
+    for (size_t i = 0; i < s.size(); i++) {
       while (j < s.size() && (count1 * count2 * count3 == 0)) {
         if (s[j] == 'a') count1++;
         else if (s[j] == 'b')
